fix: router content inside note: this also did a major fix on all packages regarding scripts note: router also expeses window.location with params/route key issue: #38, #54 changelog: - add: new page component : showcase
- add: new page component : showcase

- add: new page component : showcase

- add: uuid and DOMpath to CustomElement

- fix: router inside web-component + document proxy + referer in fetch

- fix: server scripts missing propper location to load + silencing kill output + safer kill and delete of temp folder

- fix: offline support by calling server with workspace flag

- fix: rootdir reference cleanup

- fix: sass shortened

- fix: start script cleanup

- fix: cleanup package.jsons

- fix: remove init file

- fix: correcting name of core

- clean: component script

- clean: build scripts

- fix: theme should also copy over git repo

- clean: server cleanup (one step closer to remove root color for unix system)

- fix: using new root method

- fix: calling server from root

- fix: component script can now be called from root instead of local file

- clean: build dont need to export rootdir

- fix: remane to root

- fix: exposing location proxy for params and route variables (same)

- fix: unit test for router to check new window.location params

- patch: server

- fix: using no sig mode on start and local test

+++ packages/system/server/src/run.sh
@@ -1,11 +1,8 @@
 #!/bin/bash
 
-# variables
 source .config
 export SCRIPTDIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
-export ROOTDIR=$(realpath $ROOTDIR_RELATIVE)
-source "$ROOTDIR/.config"
-loadcolors
+export ROOTDIR=$(npm prefix)
 
 # extracting flags
 export LIVE=false
@@ -25,6 +22,8 @@ for arg in "$@"; do
     export LOCATION="${arg#*=}"
   elif [[ $arg == --open ]]; then 
     export DOOPEN=true
+  elif [[ $arg == --nosig ]]; then 
+    NOSIG=true
   
   # Log Levels 
   elif [[ $arg == --log-level=* ]]; then 
@@ -152,13 +151,22 @@ function cleanup() {
   fi 
 
   # run all necessary cleanups 
-  kill $server_pid
-  rm -rf "$LOCATION/.temp" 
+  if [[ -n "$server_pid" ]]; then
+    kill $server_pid > /dev/null
+  fi
+  
+  if [[ -d "$LOCATION/.temp" ]]; then 
+    rm -rf "$LOCATION/.temp" 
+  fi
   echo ""
 }
 
-# Trap to call cleanup function when the script receives a SIGINT (Ctrl+C)
-trap cleanup SIGINT EXIT SIGTERM
+if [[ $NOSIG == true ]]; then 
+  trap cleanup EXIT
+else 
+  # Trap to call cleanup function when the script receives a SIGINT (Ctrl+C)
+  trap cleanup SIGINT EXIT SIGTERM
+fi 
 
 if [[ -f "$SCRIPTDIR/bundle.js" ]]; then 
   node "$SCRIPTDIR/bundle.js" &
